scripts/requirements.txt
google-cloud-compute
google-cloud-pubsub